Adding test and DECONV

+++ generate_data/download.sh
@@ -1,18 +1,21 @@
 # CelebA images
-URL=https://www.dropbox.com/s/3e5cmqgplchz85o/CelebA_nocrop.zip?dl=0
-ZIP_FILE=./data/CelebA_nocrop.zip
-mkdir -p ./data/
-wget -N $URL -O $ZIP_FILE
-unzip $ZIP_FILE -d ./data/CelebA
-rm $ZIP_FILE
-mkdir -p ./data/CelebA/Org
-mv ./data/CelebA/CelebA_nocrop/images ./data/CelebA/Org
-#find ./data/CelebA/CelebA_nocrop/images -exec mv {} ./data/CelebA/Org \;
-rm -r ./data/CelebA/CelebA_nocrop
+if ! [ -x "$(command -v kaggle)" ]; then
+  echo "Make sure you have kaggle installed (pip install kaggle) and kaggle API properly located at $HOME/.kaggle/kaggle.json."
+  exit 1
+fi
 
-# CelebA attribute labels
-URL=https://www.dropbox.com/s/auexdy98c6g7y25/list_attr_celeba.zip?dl=0
-ZIP_FILE=./data/list_attr_celeba.zip
-wget -N $URL -O $ZIP_FILE
-unzip $ZIP_FILE -d ./data/CelebA
+kaggle datasets download -d jessicali9530/celeba-dataset
+ZIP_FILE=celeba-dataset.zip
+FOLDER=./data/CelebA
+mkdir -p $FOLDER
+unzip $ZIP_FILE -d $FOLDER
 rm $ZIP_FILE
+
+mv $FOLDER/list_eval_partition.csv $FOLDER/train_val_test.txt #Train/Val/Test Partition
+mv $FOLDER/list_attr_celeba.csv $FOLDER/list_attr_celeba.txt #List Attributes
+
+# CelebA aligned
+unzip $FOLDER/img_align_celeba.zip -d $FOLDER
+rm $FOLDER/img_align_celeba.zip
+rm $FOLDER/*.csv
+
